ai-scheduler/requirements.txt
Flask==3.1.1
Flask-APScheduler==1.12.4
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
openai==1.95.1
python-dotenv==1.0.1
pytz
